refactor(regular_expression): extract `has_unsupported_regular_expression_pattern` (#13468)
Extracted the unsupported regex syntax code from the transformer to oxc_regular_expression to use it from the minifier.
