+++ crates/oxc_regular_expression/src/ast_impl/mod.rs
@@ -1,4 +1,5 @@
 mod allocator;
 mod display;
 mod span;
+pub mod support;
 pub mod visit;
