+++ crates/oxc_regular_expression/src/lib.rs
@@ -15,6 +15,7 @@ mod generated {
 
 pub mod ast;
 pub use crate::{
+    ast_impl::support::{RegexUnsupportedPatterns, has_unsupported_regular_expression_pattern},
     ast_impl::visit,
     options::Options,
     parser::{ConstructorParser, LiteralParser},
